requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect==1.1.0
Django==3.2.14
django-environ==0.9.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.7.1
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.0
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.3
parsel==1.6.0
Pillow==9.2.0
Protego==0.2.1
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pygsheets==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyPdf==1.13
PyPDF2==2.4.2
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
qrcode==7.3.1
queuelib==1.6.2
reportlab==3.6.10
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.0
Twisted==22.4.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
w3lib==1.22.0
webencodings==0.5.1
whitenoise==6.2.0
zope.interface==5.4.0
